Divide backend client hierarchy to separate modules.
